fix: PostgreSQL兼容性修复
- Database.execute()添加自动commit机制
- 处理CREATE/DROP等无结果语句（cursor.description为None）
- 所有schema文件的DOUBLE类型改为DOUBLE PRECISION
- PostgreSQL连接测试和初始化验证成功

数据库状态：
- 28张表全部创建成功
- 游标表初始化27条记录

+++ database/schemas/balancesheet_schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS balancesheet (
     report_type VARCHAR(20),  -- 报表类型
     comp_type VARCHAR(20),  -- 公司类型(1一般工商业2银行3保险4证券)
     end_type VARCHAR(20),  -- 报告期类型
-    total_share DOUBLE,  -- 期末总股本
+    total_share DOUBLE PRECISION,  -- 期末总股本
     cap_rese REAL,  -- 资本公积金
     undistr_porfit REAL,  -- 未分配利润
     surplus_rese REAL,  -- 盈余公积金
